ml-dl-project/starters/cv-project-starter/cv-project/requirements.txt
torch
torchvision
torchmetrics==1.4.0
pyyaml
tqdm
matplotlib
seaborn
scikit-learn
